move template tests to load_file tests.
