--- api/pkg/template/kubernetes/fixtures/invalid.yml
@@ -1,2 +1,0 @@
----
-name: name: {{.Name}}
