--- api/pkg/template/kubernetes/fixtures/simple.yml
@@ -1,2 +1,0 @@
----
-name: {{.Name}}
